Add input, output node colors
