+++ Scripts/Utils/CapricornColors.cs
@@ -5,7 +5,9 @@ namespace Dunward.Capricorn
 {
     public static class CapricornColors
     {
-        public static Color Green => new Color(0.69f, 0.98f, 0.34f);
+        public static Color InputTitleHeader => Color.green;
+        public static Color OutputTitleHeader => Color.red;
+        public static Color Port => new Color(0.69f, 0.98f, 0.34f);
     }
 }
 #endif
